+++ packages/frontend/apps/ios/App/Packages/Intelligents/Sources/Intelligents/Interface/View/InputBox/ViewModel/InputBoxViewModel.swift
@@ -66,7 +66,7 @@ public class InputBoxViewModel: ObservableObject {
       .assign(to: \.canSend, on: self)
       .store(in: &cancellables)
   }
-  
+
   public func clearAllAttachments() {
     imageAttachments.removeAll()
     fileAttachments.removeAll()
